# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proyecta360 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/risks": {
      "post": {
        "summary": "Create Risk",
        "operationId": "create_risk_api_risks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Risk Api Risks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources": {
      "post": {
        "summary": "Create Resource",
        "operationId": "create_resource_api_resources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Resource Api Resources Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/components": {
      "post": {
        "summary": "Create Component",
        "operationId": "create_component_api_components_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComponentIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Component Api Components Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations": {
      "post": {
        "summary": "Create Conversation",
        "operationId": "create_conversation_api_conversations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationThreadIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Conversation Api Conversations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/conversations/{thread_id}/messages": {
      "post": {
        "summary": "Create Conversation Message",
        "operationId": "create_conversation_message_api_conversations__thread_id__messages_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationMessageIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Conversation Message Api Conversations  Thread Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComponentIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "methodology": {
            "type": "string",
            "title": "Methodology",
            "default": "Hibrida"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          },
          "objective": {
            "type": "string",
            "title": "Objective",
            "default": ""
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name"
        ],
        "title": "ComponentIn"
      },
      "ConversationMessageIn": {
        "properties": {
          "thread_id": {
            "type": "integer",
            "title": "Thread Id"
          },
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "message": {
            "type": "string",
            "maxLength": 2500,
            "minLength": 1,
            "title": "Message"
          },
          "mentions": {
            "type": "string",
            "title": "Mentions",
            "default": ""
          },
          "evidence_url": {
            "type": "string",
            "title": "Evidence Url",
            "default": ""
          },
          "message_type": {
            "type": "string",
            "title": "Message Type",
            "default": "Comentario"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "project_id",
          "message"
        ],
        "title": "ConversationMessageIn"
      },
      "ConversationThreadIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "maxLength": 180,
            "minLength": 1,
            "title": "Title"
          },
          "context_type": {
            "type": "string",
            "title": "Context Type",
            "default": "Proyecto"
          },
          "context_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Id"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": "Seguimiento"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Abierta"
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title"
        ],
        "title": "ConversationThreadIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResourceIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "email": {
            "type": "string",
            "title": "Email",
            "default": ""
          },
          "capacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Capacity",
            "default": 100
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "name"
        ],
        "title": "ResourceIn"
      },
      "RiskIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 1,
            "title": "Title"
          },
          "probability": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Probability",
            "default": 1
          },
          "impact": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Impact",
            "default": 1
          },
          "response": {
            "type": "string",
            "title": "Response",
            "default": ""
          },
          "mitigation_plan": {
            "type": "string",
            "title": "Mitigation Plan",
            "default": ""
          },
          "contingency_plan": {
            "type": "string",
            "title": "Contingency Plan",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Abierto"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title"
        ],
        "title": "RiskIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}